refactor: 페이지별 네이밍 변경
