+++ src/pages/PostDetail/style.ts
