+++ src/pages/Profile/index.tsx
@@ -1,4 +1,4 @@
-import { UserInfoWrapper, RightButton } from '@/pages/UserPage/style';
+import { UserInfoWrapper, RightButton } from '@/pages/Profile/style';
 import FollowButton from '@/components/User/FollowButton';
 import UserPostList from '@/components/User/UserPostList';
 import UserInfo from '@/components/User/UserInfo';
