+++ src/pages/ProfileEdit/index.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { Outlet, useNavigate, useParams } from 'react-router-dom';
 
-import { Container, CardWrapper } from '@/pages/SettingPage/style';
+import { Container, CardWrapper } from '@/pages/ProfileEdit/style';
 import { getUser } from '@/slices/user/thunk';
 import { useDispatch } from '@/store';
 import { Avatar } from '@/components/common';
