+++ src/pages/Sign/index.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { Container, Card, SignText, GoHome } from '@/pages/SignPage/style';
+import { Container, Card, SignText, GoHome } from '@/pages/Sign/style';
 import { Text } from '@/components/common';
 import In from '@/components/Sign/In';
 import Up from '@/components/Sign/Up';
